pyproject.toml
[project]
name = "rffl-tools"
version = "0.2.0"
description = "RFFL Fantasy Football data toolkit - export, validate, and automate"
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "Kyle Thorsen"}]

dependencies = [
    # Core - ESPN API and data processing
    "espn_api>=0.39.1",
    "pandas>=2.2.0",
    "python-dateutil>=2.9.0",
    "requests>=2.31.0",
    
    # CLI and UI
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    
    # Configuration and validation
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
rffl = "rffl.cli:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

